// src/models/SchoolModel.ts
interface SchoolModel {
  AccountCode: string;
  AccountID: string;
  AccountName: string;
  Address1: string;
  Address2: string;
  City: string;
  Country: string;
  CreatedBy: string;
  CreatedOn: string;
  SchoolCode: string;
  SchoolID: number;
  SchoolName: string;
  State: string;
  Status: string;
  UpdatedBy: string;
  UpdatedOn: string;
  ZipCode: string;
}
export type { SchoolModel };